requirements.txt
streamlit==1.37.1
pandas==2.2.2
openpyxl==3.1.2
requests==2.32.3
beautifulsoup4==4.12.3
Pillow==10.4.0
ImageHash==4.3.1
serpapi==0.1.5
